новый аргумент double для больших чисел

+++ Program.cs
@@ -1,5 +1,5 @@
 ﻿// Пишем прогу для расчета факториала новым способом
- int Factorial (int n)
+ double Factorial (int n)
  {
     // 1! = 1
     // 0! = 1
